requirements.txt
discord.py
python-dotenv
firebase-admin
fastapi
uvicorn
requests
